requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3

# Object detection
ultralytics==8.0.196

# OCR
easyocr==1.7.1

# Text-to-speech
pyttsx3==2.90
pydub==0.25.1

# Utilities
pillow==10.0.1